requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
python-dotenv==1.0.1
ultralytics==8.2.32
torch==2.3.1
torchvision==0.18.1
timm==1.0.7
opencv-python-headless==4.10.0.84
python-multipart==0.0.9
